Add verbose option for filesystem builds. The default is not-verbose.

+++ build.sh
@@ -18,13 +18,14 @@ Usage() {
     echo -e "\t\t-p --path (optional) Specify a path to download and build the sources."
     echo -e "\t\t-n --noconfirm (optional) Build systems without confirmation."
     echo -e "\t\t-i --install-dep (optional) Attempt to install dependencies."
+    echo -e "\t\t-v --verbose (optional) Show make output on screen for filesystem builds as well as write it to the log file."
     echo -e "\t\t-h --help -? Display this message."
     exit 0
 }
 [[ -n "$arch" ]] && unset "$arch"
 
-shortopts="?hkfnia:p:"
-longopts="help,kernel-only,filesystem-only,noconfirm,install-dep,arch:,path:"
+shortopts="?hkfnia:p:v"
+longopts="help,kernel-only,filesystem-only,noconfirm,install-dep,arch:,path:,verbose"
 
 optargs=$(getopt -o "$shortopts" -l "$longopts" -n "$0" -- "$@")
 [[ $? -ne 0 ]] && Usage
@@ -50,6 +51,10 @@ while :; do
             ;;
         -i | --install-dep)
             installDep="y"
+            shift
+            ;;
+        -v | --verbose)
+            verbose="y"
             shift
             ;;
         -a | --arch)
@@ -80,6 +85,7 @@ done
 [[ -z $buildPath ]] && buildPath="$(dirname "$(readlink -f "$0")")"
 [[ -z $confirm ]] && confirm="y"
 [[ -z $installDep ]] && installDep="n"
+[[ -z $verbose ]] && verbose="n"
 
 checkDependencies
 installDependencies "$installDep"
@@ -184,27 +190,50 @@ function buildFilesystem() {
             return
         fi
     fi
-    bash -c "while true; do echo \$(date) - building ...; sleep 30s; done" &
-    PING_LOOP_PID=$!
-    case "${arch}" in
-        x64)
-            make > "buildroot$arch.log" 2>&1
-            status=$?
-            ;;
-        x86)
-            make ARCH=i486 > "buildroot$arch.log" 2>&1
-            status=$?
-            ;;
-        arm64)
-            make ARCH=aarch64 CROSS_COMPILE=aarch64-linux-gnu- > "buildroot$arch.log" 2>&1
-            status=$?
-            ;;
-        *)
-            make > "buildroot$arch.log" 2>&1
-            status=$?
-            ;;
-    esac
-    kill $PING_LOOP_PID
+
+    if [[ $verbose == "y" ]]; then
+        case "${arch}" in
+            x64)
+                make | tee "buildroot$arch.log"
+                status=${PIPESTATUS[0]}
+                ;;
+            x86)
+                make ARCH=i486 | tee "buildroot$arch.log"
+                status=${PIPESTATUS[0]}
+                ;;
+            arm64)
+                make ARCH=aarch64 CROSS_COMPILE=aarch64-linux-gnu- | tee "buildroot$arch.log"
+                status=${PIPESTATUS[0]}
+                ;;
+            *)
+                make | tee "buildroot$arch.log"
+                status=${PIPESTATUS[0]}
+                ;;
+        esac
+    else
+        bash -c "while true; do echo \$(date) - building ...; sleep 30s; done" &
+        PING_LOOP_PID=$!
+        case "${arch}" in
+            x64)
+                make > "buildroot$arch.log" 2>&1
+                status=$?
+                ;;
+            x86)
+                make ARCH=i486 > "buildroot$arch.log" 2>&1
+                status=$?
+                ;;
+            arm64)
+                make ARCH=aarch64 CROSS_COMPILE=aarch64-linux-gnu- > "buildroot$arch.log" 2>&1
+                status=$?
+                ;;
+            *)
+                make > "buildroot$arch.log" 2>&1
+                status=$?
+                ;;
+        esac
+        kill $PING_LOOP_PID
+    fi
+
     [[ $status -gt 0 ]] && tail "buildroot$arch.log" && exit $status
     cd ..
     [[ ! -d dist ]] && mkdir dist
